requirements.txt
pyrealsense2
numpy
opencv-python
ultralytics
Pillow
onnxruntime
python-dotenv